OTC-289: Rename submodules
