+++ policy_notification/NotificationTemplates/__init__.py
@@ -1,0 +1,1 @@
+from policy_notification.NotificationTemplates.DefaultSMSTemplates import DefaultSMSTemplates
