--- policy_notification/SMSTemplates/__init__.py
@@ -1,1 +1,0 @@
-from policy_notification.SMSTemplates.DefaultSMSTemplates import DefaultSMSTemplates
